Refactored tests for App.
